feature: add authorization to the app

+++ backend/routes/posts.js
@@ -35,7 +35,8 @@ router.post("",
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     console.log(createdPost);
@@ -62,13 +63,18 @@ router.put("/:id",
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
   console.log(post);
-  Post.updateOne({_id: req.params.id}, post).then(result=>{
-    console.log(result);
-    res.status(200).json({
-      message: 'Post updated successfully'
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post).then(result=>{
+    if (result.nModified > 0) {
+      res.status(200).json({
+        message: 'Post updated successfully'
+      });
+    }
+    res.status(401).json({
+      message: 'You are not authorized to update this post!'
     });
   });
 });
@@ -111,9 +117,15 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post delete!" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({
+        message: 'Post delete successfully!'
+      });
+    }
+    res.status(401).json({
+      message: 'You are not authorized to delete this post!'
+    });
   });
 
 });
